refactor: binance and bitstamp ws synchronization
After discovering that the rest request can have a timestamp of a second or more
prior to the latest websocket event timestamps, now sleep a short amount of time
after channel subscription before requesting the initial orderbook snapshot to give.
the websocket buffer a chance to populate with messages which are older than the snapshot.

Very careful checks that the event buffer is sufficient to synchronize our initial local state
with the remote orderbook state.