refactor: websocket clients no longer keep state
Prior to now a websocket client would maintain its own orderbook, updaing on snapshots
or deltas. Now websocket clients' just convert incoming external snapshot / delta
messages into an internal message protocol and forward downstream.

A benefit of this is it saves memory and cpu as downstream consumers are likely to
construct their own (copy) of an order book anyway (for example, an aggregator).