fix: remove Mutex in delta clients
Prior to this the orderbook field of a delta client was wrapped in a
Mutex so it was Sync, and could be held across await points. This was
unnecessary as there was no contention on the object. Instead now
the receiver of WebsocketClient::manage_connection is wrapped in an
Arc, meaning all fields are automatically Sync and there is no Mutex.